refactor(domain): move app/types into app/core/domain/types
Consolidate shared type contracts under the domain package and remove the
top-level app/types package. Inline unused chat protocol types into
chat_sdk_adapter since that module is the only consumer today.
