+++ TOOL_INTEGRATION_CHECKLIST.md
@@ -86,7 +86,7 @@ If the tool/integration is relevant to investigations:
 
 - [ ] Review alert-source seeding in `app/core/domain/alerts/alert_source.py`
 - [ ] Review source-priority/prompt mapping in `app/core/orchestration/node/investigate/prompt.py`
-- [ ] Review evidence/source registration in `app/types/` or related state models when relevant
+- [ ] Review evidence/source registration in `app/core/domain/types/` or related state models when relevant
 - [ ] Add scenario coverage proving the tool surfaces useful RCA evidence
 
 If the integration is first-class for an `alert_source`, the source-to-tool maps must be reviewed explicitly.
