+++ tests/core/domain/types/test_retrieval_plan.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from pydantic import BaseModel, Field
 
-from app.types.retrieval import (
+from app.core.domain.types.retrieval import (
     AggregationSpec,
     RetrievalControlsMap,
     RetrievalIntent,
